Prevent double initialization of the same game
Sometimes there is a double-read of moves, which looks like it arises
from #board-liveGame-<game> node getting inserted twice (perhaps with it
removed in between?).

Use a simple fix (checking that whether we are already watching the same
game), which also allows us to switch observing of a new game.